fix: log the client IP if X-Real-IP header is being used
Currently the library attempts to log IP addresses using the `X-Forwarded-For` header but this results in the proxy IP being logged as well. The downside of this is if you are using something like nginx to reverse proxy all traffic to the local services then you end up in a situation where every user registering/accessing the site logs the proxy address and that results in the `max.connections.per.ip` setting being applied globally i.e. if `max.connections.per.ip` is set to 2 then only 2 users can register in the hotel.